refactor(runner): split codex_app_server.go into per-concern files (#410)
codex_app_server.go had grown to 1577 lines: subprocess command construction,
the JSON-RPC transport/protocol client, turn-completion handling, runtime-event
recording, and the approval/dynamic-tool bridge all in one file. Split it into
five per-concern files in the same package runner, moving whole declarations
verbatim (no logic change):

  codex_app_server.go (587)           runner entry, command/client core, the
                                      JSON-RPC transport (run/request/read*),
                                      and small prompt/artifact helpers
  codex_app_server_cmd.go (154)        command construction
                                      (split/shell-syntax checks)
  codex_app_server_turn.go (365)      turn-completion await + failure/quota
                                      classification + retry-after parsing
  codex_app_server_events.go (155)    SPEC runtime-event recording + payload
                                      snake_case normalization
  codex_app_server_approval.go (322)  server->client request answering:
                                      approval-policy gate + dynamic tool bridge

Pure structural move — no logic change. The oversized functions
(awaitTurnCompletion, run, Run, handleDynamicToolCall, ...) are moved verbatim;
their internal decomposition touches behavior on the protocol path and is
deferred to a separate, characterization-test-backed PR.

Verified behavior-preserving:
- top-level declaration set identical before/after (76 == 76);
- go test -race ./internal/runner/ passes (excluding the pre-existing
  environment-only TestCodexAppServerInitializeTimeoutDiagnostics, which fails
  identically on main under the local sandbox and passes in CI);
- funlen/gocognit finding set byte-identical to main (11 == 11, no new
  violations);
- gofmt/go vet/build clean.

Addresses #410 (codex_app_server.go file-size hotspot).
